verion final 01/05/2024

+++ app/Http/Controllers/Stock.php
@@ -11,13 +11,19 @@ class Stock extends Controller
 {
     public function home()
     {
-        $produits = Produits::all();
+        $produits = Produits::paginate(6);
         return view('home', compact('produits'));
     }
     public function Fournisseur()
     {
-        $fournisseurs = Fournisseurs::all();
+        $fournisseurs = Fournisseurs::paginate(6);
+        ;
         return view('fournisseur', compact('fournisseurs'));
+    }
+    public function Catégorie()
+    {
+        $catégories = Categories::paginate(6);
+        return view('catégorie', compact('catégories'));
     }
     public function Produits()
     {
@@ -25,28 +31,36 @@ class Stock extends Controller
         $catégories = Categories::all();
         return view('produits', compact('catégories', 'fournisseurs'));
     }
-    public function Catégorie()
-    {
-        $catégories = Categories::all();
-        return view('catégorie', compact('catégories'));
-    }
     public function RemoveCat($id)
     {
-        $catégorie = Categories::find($id);
-        $catégorie->delete();
-        return redirect('/Catégorie');
+        $categorie = Categories::find($id);
+        if (!$categorie) {
+            return redirect('/Catégorie')->with('error', 'Catégorie non trouvée.');
+        }
+
+        $categorie->delete();
+        return redirect('/Catégorie')->with('success', 'La catégorie a été supprimée avec succès.');
     }
+
     public function RemoveFour($id)
     {
         $fournisseur = Fournisseurs::find($id);
+        if (!$fournisseur) {
+            return redirect('/fournisseur')->with('error', 'Fournisseur non trouvé.');
+        }
+
         $fournisseur->delete();
-        return redirect('/fournisseur');
+        return redirect('/fournisseur')->with('success', 'Le fournisseur a été supprimé avec succès.');
     }
     public function RemoveProd($id)
     {
         $produit = Produits::find($id);
+        if (!$produit) {
+            return redirect('/')->with('error', 'Produit non trouvé.');
+        }
+
         $produit->delete();
-        return redirect('/');
+        return redirect('/')->with('success', 'Le produit a été supprimé avec succès.');
     }
     public function search(Request $request)
     {
@@ -86,14 +100,41 @@ class Stock extends Controller
                 'price' => $request->get('price'),
                 'quantité' => $request->get('quantité'),
                 'fournisseur_id' => $request->get('fournisseur_id'),
-                'categorie_id' => $request->get('categorie_id'),
+                'categorie_id' => $request->get('categorie_id')
             ]);
 
             session()->flash('success', 'Produit ajouté avec succès.');
         } catch (\Exception $e) {
-            session()->flash('error', 'Une erreur est survenue lors de l\'ajout du produit.');
+            session()->flash('error', 'Une erreur est survenue lors de l\'ajout du produit. Assurez-vous de remplir tous les champs');
         }
 
+        return redirect('/');
+    }
+    public function update($id)
+    {
+        $produit = Produits::find($id);
+        $fournisseurs = Fournisseurs::all();
+        $catégories = Categories::all();
+        return view('update', compact('produit', 'catégories', 'fournisseurs'));
+    }
+    public function update_produit(Request $request)
+    {
+        $produit = Produits::find($request->get('id'));
+        try {
+            $produit->update(
+                [
+                    'name' => $request->get('name'),
+                    'description' => $request->get('description'),
+                    'price' => $request->get('price'),
+                    'quantité' => $request->get('quantité'),
+                    'fournisseur_id' => $request->get('fournisseur_id'),
+                    'categorie_id' => $request->get('categorie_id')
+                ]
+            );
+            session()->flash('success', 'Produit Modifié avec succès.');
+        } catch (\Exception $e) {
+            session()->flash('error', 'Une erreur est survenue lors de Modifieé du produit.');
+        }
         return redirect('/');
     }
 }
